+++ controllers/publishersController.cjs
@@ -1,5 +1,9 @@
 
+const { validationResult, matchedData } = require('express-validator');
 const publisherDB = require('../db/publishers/repository.cjs');
+
+const nameValidator = require('../middlewares/validation/formValidation.cjs').body.nameValidator;
+const idValidator = require('../middlewares/validation/formValidation.cjs').params.idValidator;
 
 async function getAllPublishers(req, res) {
     const publishers = await publisherDB.getAllPublishers();
@@ -11,33 +15,49 @@ function getCreateForm(req, res) {
 }
 
 async function postCreateForm(req, res) {
-    const {name} = req.body;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).render('pages/publishers/createForm', { errors: errors.array() });
+    }
+    const {name} = matchedData(req);
     await publisherDB.addPublisher(name);
     res.redirect("/publishers");
 }
 
 async function postDeletePublisher(req, res) {
-    const {id} = req.params;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.sendStatus(400);
+    }
+    const {id} = matchedData(req);
     await publisherDB.deletePublisher(id);
     res.redirect("/publishers");
 }
 
 async function getUpdateForm(req, res) {
-    const {id} = req.params;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.sendStatus(400);
+    }
+    const {id} = matchedData(req);
     const publisher = await publisherDB.getPublisherById(id);
     res.render('pages/publishers/updateForm', {publisher});
 }
 
 async function postUpdatePublisher(req, res) {
-    const {id} = req.params;
-    const {name} = req.body;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).render('pages/publishers/updateForm', { errors: errors.array(), publisher: {id: req.params.id, name: req.body.name} });
+    }
+    const {id} = matchedData(req, {locations: ['params']});
+    const {name} = matchedData(req, {locations: ['body']});
     await publisherDB.updatePublisher(id, name);
     res.redirect("/publishers");
 }
 
 module.exports.getAllPublishers = getAllPublishers;
 module.exports.getCreateForm = getCreateForm;
-module.exports.postCreateForm = postCreateForm;
-module.exports.postDeletePublisher = postDeletePublisher;
-module.exports.getUpdateForm = getUpdateForm;
-module.exports.postUpdatePublisher = postUpdatePublisher;
+module.exports.postCreateForm = [nameValidator, postCreateForm];
+module.exports.postDeletePublisher = [idValidator, postDeletePublisher];
+module.exports.getUpdateForm = [idValidator, getUpdateForm];
+module.exports.postUpdatePublisher = [idValidator, nameValidator, postUpdatePublisher];
