+++ controllers/regionsController.cjs
@@ -1,5 +1,9 @@
 
+const { validationResult, matchedData } = require('express-validator');
 const regionDB = require('../db/regions/repository.cjs');
+
+const nameValidator = require('../middlewares/validation/formValidation.cjs').body.nameValidator;
+const idValidator = require('../middlewares/validation/formValidation.cjs').params.idValidator;
 
 async function getAllRegions(req, res) {
     const regions = await regionDB.getAllRegions();
@@ -11,33 +15,49 @@ function getCreateForm(req, res) {
 }
 
 async function postCreateForm(req, res) {
-    const {name} = req.body;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).render('pages/regions/createForm', { errors: errors.array() });
+    }
+    const {name} = matchedData(req);
     await regionDB.addRegion(name);
     res.redirect("/regions");
 }
 
 async function postDeleteRegion(req, res) {
-    const {id} = req.params;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.sendStatus(400);
+    }
+    const {id} = matchedData(req);
     await regionDB.deleteRegion(id);
     res.redirect("/regions");
 }
 
 async function getUpdateForm(req, res) {
-    const {id} = req.params;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.sendStatus(400);
+    }
+    const {id} = matchedData(req);
     const region = await regionDB.getRegionById(id);
     res.render('pages/regions/updateForm', {region});
 }
 
 async function postUpdateRegion(req, res) {
-    const {id} = req.params;
-    const {name} = req.body;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).render('pages/regions/updateForm', { errors: errors.array(), region: {id: req.params.id, name: req.body.name} });
+    }
+    const {id} = matchedData(req, {locations: ['params']});
+    const {name} = matchedData(req, {locations: ['body']});
     await regionDB.updateRegion(id, name);
     res.redirect("/regions");
 }
 
 module.exports.getAllRegions = getAllRegions;
 module.exports.getCreateForm = getCreateForm;
-module.exports.postCreateForm = postCreateForm;
-module.exports.postDeleteRegion = postDeleteRegion;
-module.exports.getUpdateForm = getUpdateForm;
-module.exports.postUpdateRegion = postUpdateRegion;
+module.exports.postCreateForm = [nameValidator, postCreateForm];
+module.exports.postDeleteRegion = [idValidator, postDeleteRegion];
+module.exports.getUpdateForm = [idValidator, getUpdateForm];
+module.exports.postUpdateRegion = [idValidator, nameValidator, postUpdateRegion];
